+++ pkg/storage/bucket/s3/config_test.go
@@ -7,12 +7,14 @@ package s3
 
 import (
 	"encoding/base64"
-	"net/http"
-	"testing"
 
+	s3_service "github.com/aws/aws-sdk-go/service/s3"
 	"github.com/grafana/dskit/flagext"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+
+	"net/http"
+	"testing"
 )
 
 func TestSSEConfig_Validate(t *testing.T) {
@@ -76,7 +78,7 @@ func TestConfig_Validate(t *testing.T) {
 					BucketName:       "mimir-block",
 					SSE:              *sseCfg,
 					SignatureVersion: SignatureVersionV4,
-					StorageClass:     StorageClassStandard,
+					StorageClass:     s3_service.StorageClassStandard,
 				}
 				return cfg
 			},
@@ -93,7 +95,7 @@ func TestConfig_Validate(t *testing.T) {
 		"should pass if valid storage signature version is set": {
 			setup: func() *Config {
 				return &Config{
-					SignatureVersion: SignatureVersionV4, StorageClass: StorageClassStandardInfrequentAccess,
+					SignatureVersion: SignatureVersionV4, StorageClass: s3_service.StorageClassStandard,
 				}
 			},
 		},
@@ -103,7 +105,7 @@ func TestConfig_Validate(t *testing.T) {
 					Endpoint:         "mimir-blocks.s3.eu-central-1.amazonaws.com",
 					BucketName:       "mimir-blocks",
 					SignatureVersion: SignatureVersionV4,
-					StorageClass:     StorageClassGlacier,
+					StorageClass:     s3_service.StorageClassStandard,
 				}
 			},
 			expected: errInvalidEndpointPrefix,